+++ src/main/java/com/tiph/dysonsphereproject/common/blocks/entities/DysonEnergyBlockEntity.java
@@ -1,18 +1,45 @@
 package com.tiph.dysonsphereproject.common.blocks.entities;
 
 import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockState;
+import net.neoforged.neoforge.common.capabilities.Capabilities;
+import net.neoforged.neoforge.common.capabilities.Capability;
+import net.neoforged.neoforge.common.util.LazyOptional;
 import net.neoforged.neoforge.energy.IEnergyStorage;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 public abstract class DysonEnergyBlockEntity extends DysonBlockEntity implements IEnergyStorage {
 
   private int energy;
+  private LazyOptional<IEnergyStorage> lazyEnergyHandler = LazyOptional.empty();
 
   protected DysonEnergyBlockEntity(BlockEntityType<?> entityType, BlockPos pos, BlockState state) {
     super(entityType, pos, state);
     energy = 0;
+  }
+
+  @Override
+  public @NotNull <T> LazyOptional<T> getCapability(@NotNull Capability<T> cap, @Nullable Direction side) {
+    if (cap == Capabilities.ENERGY && side != Direction.UP) {
+      return lazyEnergyHandler.cast();
+    }
+    return super.getCapability(cap, side);
+  }
+
+  @Override
+  public void onLoad() {
+    super.onLoad();
+    lazyEnergyHandler = LazyOptional.of(() -> this);
+  }
+
+  @Override
+  public void invalidateCaps() {
+    super.invalidateCaps();
+    lazyEnergyHandler.invalidate();
   }
 
   @Override
