+++ src/main/java/com/tiph/dysonsphereproject/common/blocks/entities/GroundStationBlockEntity.java
@@ -3,16 +3,9 @@ package com.tiph.dysonsphereproject.common.blocks.entities;
 import com.tiph.dysonsphereproject.common.init.DysonBlockEntities;
 import com.tiph.dysonsphereproject.util.OrbitalCollectorSavedData;
 import net.minecraft.core.BlockPos;
-import net.minecraft.core.Direction;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
-import net.neoforged.neoforge.common.capabilities.Capabilities;
-import net.neoforged.neoforge.common.capabilities.Capability;
-import net.neoforged.neoforge.common.util.LazyOptional;
-import net.neoforged.neoforge.energy.IEnergyStorage;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
 public class GroundStationBlockEntity extends DysonEnergyBlockEntity {
 
@@ -23,31 +16,8 @@ public class GroundStationBlockEntity extends DysonEnergyBlockEntity {
   // Could have different kinds of orbital collectors with different power gen
   private static final int COLLECTOR_POWER_GEN = 100;
 
-  private LazyOptional<IEnergyStorage> lazyEnergyHandler = LazyOptional.empty();
-
   public GroundStationBlockEntity(BlockPos pos, BlockState state) {
     super(DysonBlockEntities.GROUND_STATION_ENTITY.get(), pos, state);
-  }
-
-  @Override
-  public @NotNull <T> LazyOptional<T> getCapability(
-      @NotNull Capability<T> cap, @Nullable Direction side) {
-    if (cap == Capabilities.ENERGY && side != Direction.UP) {
-      return lazyEnergyHandler.cast();
-    }
-    return super.getCapability(cap, side);
-  }
-
-  @Override
-  public void invalidateCaps() {
-    super.invalidateCaps();
-    lazyEnergyHandler.invalidate();
-  }
-
-  @Override
-  public void onLoad() {
-    super.onLoad();
-    lazyEnergyHandler = LazyOptional.of(() -> this);
   }
 
   public void tick(final Level level, final BlockPos pos, final BlockState blockState) {
