+++ src/main/java/com/tiph/dysonsphereproject/common/blocks/entities/SolarGeneratorBlockEntity.java
@@ -2,7 +2,6 @@ package com.tiph.dysonsphereproject.common.blocks.entities;
 
 import com.tiph.dysonsphereproject.common.init.DysonBlockEntities;
 import net.minecraft.core.BlockPos;
-import net.minecraft.core.Direction;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.MenuProvider;
 import net.minecraft.world.entity.player.Inventory;
@@ -10,8 +9,6 @@ import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.AbstractContainerMenu;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
-import net.neoforged.neoforge.common.capabilities.Capabilities;
-import net.neoforged.neoforge.common.capabilities.Capability;
 import net.neoforged.neoforge.common.util.LazyOptional;
 import net.neoforged.neoforge.energy.IEnergyStorage;
 import org.jetbrains.annotations.NotNull;
@@ -27,27 +24,6 @@ public class SolarGeneratorBlockEntity extends DysonEnergyBlockEntity implements
 
   public SolarGeneratorBlockEntity(BlockPos pos, BlockState state) {
     super(DysonBlockEntities.SOLAR_GENERATOR_ENTITY.get(), pos, state);
-  }
-
-  @Override
-  public @NotNull <T> LazyOptional<T> getCapability(
-      @NotNull Capability<T> cap, @Nullable Direction side) {
-    if (cap == Capabilities.ENERGY && side != Direction.UP) {
-      return lazyEnergyHandler.cast();
-    }
-    return super.getCapability(cap, side);
-  }
-
-  @Override
-  public void onLoad() {
-    super.onLoad();
-    lazyEnergyHandler = LazyOptional.of(() -> this);
-  }
-
-  @Override
-  public void invalidateCaps() {
-    super.invalidateCaps();
-    lazyEnergyHandler.invalidate();
   }
 
   @Override
